put discrete/fraction allocmap to common packages
