+++ src/ai/backend/agent/resources.py
@@ -225,3 +225,124 @@ def bitmask2set(mask):
         bpos += 1
     return frozenset(bset)
     
+
+class DiscretePropertyAllocMap(AbstractAllocMap):
+    '''
+    An allocation map using discrete property.
+    The user must pass a "property function" which returns a desired resource
+    property from the device object.
+
+    e.g., 1.0 means 1 device, 2.0 means 2 devices, etc.
+    (no fractions allowed)
+    '''
+
+    def __init__(self, *args, **kwargs):
+        self.property_func = kwargs.pop('prop_func')
+        super().__init__(*args, **kwargs)
+        assert callable(self.property_func)
+        self.allocations = defaultdict(lambda: {
+            dev_id: 0 for dev_id in self.devices.keys()
+        })
+
+    def allocate(self, slots: Mapping[SlotType, Allocation], *,
+                 context_tag: str = None) -> ResourceAllocations:
+        allocation = {}
+        for slot_type, alloc in slots.items():
+            slot_allocation = {}
+            remaining_alloc = int(alloc)
+
+            # fill up starting from the most free devices
+            sorted_dev_allocs = sorted(
+                self.allocations[slot_type].items(),
+                key=lambda pair: self.property_func(self.devices[pair[0]]) - pair[1],
+                reverse=True)
+            log.debug('DiscretePropertyAllocMap: allocating {} {}',
+                      slot_type, alloc)
+            log.debug('DiscretePropertyAllocMap: current-alloc: {!r}',
+                      sorted_dev_allocs)
+
+            total_allocatable = 0
+            for dev_id, current_alloc in sorted_dev_allocs:
+                current_alloc = self.allocations[slot_type][dev_id]
+                total_allocatable += (self.property_func(self.devices[dev_id]) -
+                                      current_alloc)
+            if total_allocatable < alloc:
+                raise InsufficientResource(
+                    'DiscretePropertyAllocMap: insufficient allocatable amount!',
+                    context_tag, slot_type, str(alloc), str(total_allocatable))
+
+            slot_allocation = {}
+            for dev_id, current_alloc in sorted_dev_allocs:
+                current_alloc = self.allocations[slot_type][dev_id]
+                allocatable = (self.property_func(self.devices[dev_id]) -
+                               current_alloc)
+                if allocatable > 0:
+                    allocated = min(remaining_alloc, allocatable)
+                    slot_allocation[dev_id] = allocated
+                    self.allocations[slot_type][dev_id] += allocated
+                    remaining_alloc -= allocated
+                if remaining_alloc == 0:
+                    break
+            allocation[slot_type] = slot_allocation
+        return allocation
+
+    def free(self, existing_alloc: ResourceAllocations):
+        for slot_type, per_device_alloc in existing_alloc.items():
+            for dev_id, alloc in per_device_alloc.items():
+                self.allocations[slot_type][dev_id] -= alloc
+
+
+class FractionAllocMap(AbstractAllocMap):
+
+    def __init__(self, *args, **kwargs):
+        self.shares_per_device = kwargs.pop('shares_per_device')
+        super().__init__(*args, **kwargs)
+        self.allocations = defaultdict(lambda: {
+            dev_id: Decimal(0) for dev_id in self.devices.keys()
+        })
+
+    def allocate(self, slots: Mapping[SlotType, Allocation], *,
+                 context_tag: str = None) -> ResourceAllocations:
+        allocation = {}
+        for slot_type, alloc in slots.items():
+            slot_allocation = {}
+            remaining_alloc = Decimal(alloc).normalize()
+
+            # fill up starting from the most free devices
+            sorted_dev_allocs = sorted(
+                self.allocations[slot_type].items(),
+                key=lambda pair: self.shares_per_device[pair[0]] - pair[1],
+                reverse=True)
+            log.debug('FractionAllocMap: allocating {} {}', slot_type, alloc)
+            log.debug('FractionAllocMap: current-alloc: {!r}', sorted_dev_allocs)
+
+            total_allocatable = 0
+            for dev_id, current_alloc in sorted_dev_allocs:
+                current_alloc = self.allocations[slot_type][dev_id]
+                total_allocatable += (self.shares_per_device[dev_id] -
+                                      current_alloc)
+            if total_allocatable < alloc:
+                raise InsufficientResource(
+                    'FractionAllocMap: insufficient allocatable amount!',
+                    context_tag, slot_type, str(alloc), str(total_allocatable))
+
+            slot_allocation = {}
+            for dev_id, current_alloc in sorted_dev_allocs:
+                current_alloc = self.allocations[slot_type][dev_id]
+                allocatable = (self.shares_per_device[dev_id] -
+                               current_alloc)
+                if allocatable > 0:
+                    allocated = min(remaining_alloc, allocatable)
+                    slot_allocation[dev_id] = allocated
+                    self.allocations[slot_type][dev_id] += allocated
+                    remaining_alloc -= allocated
+                if remaining_alloc <= 0:
+                    break
+            allocation[slot_type] = slot_allocation
+        return allocation
+
+    def free(self, existing_alloc: ResourceAllocations):
+        for slot_type, per_device_alloc in existing_alloc.items():
+            for dev_id, alloc in per_device_alloc.items():
+                self.allocations[slot_type][dev_id] -= alloc
+
